refactor: outsourced Reservation validation
