+++ src/main/java/io/github/janmalch/kino/control/reservation/NewReservationControl.java
@@ -1,7 +1,6 @@
 package io.github.janmalch.kino.control.reservation;
 
 import io.github.janmalch.kino.api.model.ReservationDto;
-import io.github.janmalch.kino.api.model.SeatForPresentationDto;
 import io.github.janmalch.kino.control.Control;
 import io.github.janmalch.kino.control.ResultBuilder;
 import io.github.janmalch.kino.control.generic.NewEntityControl;
@@ -9,16 +8,15 @@ import io.github.janmalch.kino.entity.Account;
 import io.github.janmalch.kino.entity.Presentation;
 import io.github.janmalch.kino.entity.Reservation;
 import io.github.janmalch.kino.entity.Seat;
-import io.github.janmalch.kino.problem.Problem;
 import io.github.janmalch.kino.repository.Repository;
 import io.github.janmalch.kino.repository.RepositoryFactory;
 import io.github.janmalch.kino.repository.specification.AccountByEmailSpec;
 import io.github.janmalch.kino.repository.specification.Specification;
 import io.github.janmalch.kino.util.Mapping;
-import io.github.janmalch.kino.util.either.EitherResultBuilder;
-import java.util.*;
+import java.util.Date;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
-import javax.ws.rs.core.Response;
 
 public class NewReservationControl implements Control<Long> {
 
@@ -32,52 +30,14 @@ public class NewReservationControl implements Control<Long> {
 
   @Override
   public <T> T execute(ResultBuilder<T, Long> result) {
-    var validationProblem = validate();
+    var reservationValidator = new ReservationValidator();
+    var validationProblem = reservationValidator.validate(reservationDto);
     if (validationProblem.isPresent()) {
       return result.failure(validationProblem.get());
     }
 
     var mapper = new NewReservationMapper();
     return new NewEntityControl<>(reservationDto, Reservation.class, mapper).execute(result);
-  }
-
-  Optional<Problem> validate() {
-    var seatsForPresentationControl =
-        new GetSeatsWithStatusControl(reservationDto.getPresentationId());
-    var result = seatsForPresentationControl.execute(new EitherResultBuilder<>());
-    if (result.isFailure()) {
-      return Optional.of(result.getProblem());
-    }
-
-    var availableIds =
-        result
-            .getSuccess()
-            .getData()
-            .stream() // Stream mit allen Sitzen für den Saal
-            .filter(
-                seatForPresentation ->
-                    !seatForPresentation.isTaken()) // nur die verfügbaren Sitze auswählen
-            .map(SeatForPresentationDto::getId) // ID benötigt zum Vergleich
-            .collect(Collectors.toSet());
-
-    // Alle verfügbaren Sitze aus den angeforderten entfernen
-    var unavailableSeats = new HashSet<>(reservationDto.getSeatIds());
-    unavailableSeats.removeAll(availableIds);
-    // Nicht verfügbare Sitze bleiben übrig
-    if (unavailableSeats.size() > 0) {
-      var problem =
-          Problem.builder()
-              .type("new-reservation/seats-unavailable")
-              .title("Some seats in the reservation are not available")
-              .status(Response.Status.BAD_REQUEST)
-              .detail(String.format("%d seat(s) unavailable", unavailableSeats.size()))
-              .parameter("unavailable_seats", unavailableSeats)
-              .instance()
-              .build();
-      return Optional.of(problem);
-    }
-
-    return Optional.empty();
   }
 
   class NewReservationMapper implements Mapping<ReservationDto, Reservation> {
